setup first class with base model
